refactor(tests): consolidate *-inprocess into canonical test files
In-process (InMemoryTransport) tests are now the primary suites in each
file. Subprocess-only tests (SIGUSR1, kill+restart ergo, timing-sensitive
multiline boundaries) live in `(subprocess)` describe blocks in the same
file. The 6 `*-inprocess.test.ts` files are deleted.

Coverage delta: src/irc-server.ts jumps from 59%/65% (funcs/lines on
main) to 76%/87% because in-process imports createMcpServer directly,
making it visible to Bun's instrumentation. Subprocess tests spawn a
separate process so coverage never reached irc-server.ts.

Pre-existing failures tracked in #83:
- 4 unread-clearing tests (channel names contain "unread" substring)
- ml-ec6 subprocess (timing-sensitive)

1. "[Session continuation after context compaction — resume from where left off]"
   → Claude: deleted 6 old *-inprocess.test.ts files, ran coverage on
     both branches, committed everything.
