+++ test/multiline-edge-cases.test.ts
@@ -1,7 +1,6 @@
-// Tests 5+6 stay subprocess: these time out intermittently under the in-process
-// harness (timing-sensitive boundary cases). Tracked in #83.
 import { describe, it, expect, beforeAll } from 'bun:test'
 import { startErgo, isErgoAvailable, type ErgoContext } from './helpers/ergo.js'
+import { startMcpInProcess } from './helpers/mcp-inprocess.js'
 import { startMcp } from './helpers/mcp.js'
 import { connectPeer } from './helpers/peer.js'
 import { toolText } from './helpers/tool.js'
@@ -15,76 +14,86 @@ describe.if(isErgoAvailable())('multiline edge cases', () => {
   })
 
   it('empty lines preserved (consecutive \\n round-trip)', async () => {
-    const sender = await startMcp(ergo, 'ml-ec1-s')
-    const receiver = await startMcp(ergo, 'ml-ec1-r')
+    const sender = await startMcpInProcess(ergo, 'ip-ml-ec1-s')
+    const receiver = await startMcpInProcess(ergo, 'ip-ml-ec1-r')
     await Promise.all([
-      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec1' } }),
-      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec1' } }),
+      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec1' } }),
+      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec1' } }),
     ])
 
     const text = 'hello\n\nworld'
-    await sender.client.callTool({ name: 'channel_message', arguments: { channel: '#ml-ec1', text } })
+    await sender.client.callTool({ name: 'channel_message', arguments: { channel: '#ip-ml-ec1', text } })
 
     const n = await receiver.waitForNotification(
-      n => n.meta.channel === '#ml-ec1' && n.meta.sender === 'ml-ec1-s',
+      n => n.meta.channel === '#ip-ml-ec1' && n.meta.sender === 'ip-ml-ec1-s',
     )
     expect(n.content).toBe(text)
   })
 
   it('trailing newline preserved', async () => {
-    const sender = await startMcp(ergo, 'ml-ec2-s')
-    const receiver = await startMcp(ergo, 'ml-ec2-r')
+    const sender = await startMcpInProcess(ergo, 'ip-ml-ec2-s')
+    const receiver = await startMcpInProcess(ergo, 'ip-ml-ec2-r')
     await Promise.all([
-      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec2' } }),
-      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec2' } }),
+      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec2' } }),
+      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec2' } }),
     ])
 
     const text = 'hello\nworld\n'
-    await sender.client.callTool({ name: 'channel_message', arguments: { channel: '#ml-ec2', text } })
+    await sender.client.callTool({ name: 'channel_message', arguments: { channel: '#ip-ml-ec2', text } })
 
     const n = await receiver.waitForNotification(
-      n => n.meta.channel === '#ml-ec2' && n.meta.sender === 'ml-ec2-s',
+      n => n.meta.channel === '#ip-ml-ec2' && n.meta.sender === 'ip-ml-ec2-s',
     )
     expect(n.content).toBe(text)
   })
 
   it('message exactly at byte boundary takes fast path (no batch)', async () => {
-    const mcp = await startMcp(ergo, 'ml-ec3-s')
-    const peer = await connectPeer(ergo, 'ml-ec3-p')
-    await mcp.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec3' } })
-    await peer.joinChannel('#ml-ec3')
+    const mcp = await startMcpInProcess(ergo, 'ip-ml-ec3-s')
+    const peer = await connectPeer(ergo, 'ip-ml-ec3-p')
+    await mcp.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec3' } })
+    await peer.joinChannel('#ip-ml-ec3')
 
     const text = 'x'.repeat(MULTILINE_LINE_BYTES) // exactly at threshold — single PRIVMSG
     const result = await mcp.client.callTool({
       name: 'channel_message',
-      arguments: { channel: '#ml-ec3', text },
+      arguments: { channel: '#ip-ml-ec3', text },
     })
     expect(result.isError).toBeFalsy()
     expect(toolText(result)).not.toContain('batch')
 
-    await peer.waitForMessage('#ml-ec3', m => m.nick === 'ml-ec3-s' && m.text === text)
+    await peer.waitForMessage('#ip-ml-ec3', m => m.nick === 'ip-ml-ec3-s' && m.text === text)
   })
 
   it('message one byte over boundary sends as draft/multiline batch', async () => {
-    const sender = await startMcp(ergo, 'ml-ec4-s')
-    const receiver = await startMcp(ergo, 'ml-ec4-r')
+    const sender = await startMcpInProcess(ergo, 'ip-ml-ec4-s')
+    const receiver = await startMcpInProcess(ergo, 'ip-ml-ec4-r')
     await Promise.all([
-      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec4' } }),
-      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ml-ec4' } }),
+      sender.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec4' } }),
+      receiver.client.callTool({ name: 'channel_join', arguments: { channel: '#ip-ml-ec4' } }),
     ])
 
     const text = 'x'.repeat(MULTILINE_LINE_BYTES + 1) // one byte over — must split with concat
     const result = await sender.client.callTool({
       name: 'channel_message',
-      arguments: { channel: '#ml-ec4', text },
+      arguments: { channel: '#ip-ml-ec4', text },
     })
     expect(result.isError).toBeFalsy()
     expect(toolText(result)).toContain('draft/multiline batch')
 
     const n = await receiver.waitForNotification(
-      n => n.meta.channel === '#ml-ec4' && n.meta.sender === 'ml-ec4-s',
+      n => n.meta.channel === '#ip-ml-ec4' && n.meta.sender === 'ip-ml-ec4-s',
     )
     expect(n.content).toBe(text)
+  })
+})
+
+// Subprocess-only: these time out intermittently under the in-process harness
+// (timing-sensitive boundary cases). Tracked in #83.
+describe.if(isErgoAvailable())('multiline edge cases (subprocess)', () => {
+  let ergo: ErgoContext
+
+  beforeAll(async () => {
+    ergo = (await startErgo())!
   })
 
   it('exceeding max-lines: tool succeeds and something arrives', async () => {
